+++ lib/byebug/dap/commands/set_breakpoints.rb
@@ -29,9 +29,10 @@ module Byebug::DAP
         bp = find_or_add_breakpoint(verified, existing, path, rq.line)
         bp.expr = convert_breakpoint_condition(rq.condition)
         bp.hit_condition, bp.hit_value = convert_breakpoint_hit_condition(rq.hitCondition)
+        @session.set_log_point(bp, rq.logMessage)
       end
 
-      existing.each { |bp| Byebug.breakpoints.delete(bp) }
+      @session.clear_breakpoints(*existing)
 
       respond! body: {
         breakpoints: verified.map { |bp|