+++ lib/byebug/gem.rb
@@ -1,0 +1,11 @@
+module Byebug
+  module DAP
+    NAME = 'byebug-dap'
+    VERSION = '0.1.3'
+    SUMMARY = 'Debug Adapter Protocol for Byebug'
+    DESCRIPTION = 'Implements a Debug Adapter Protocol interface for Byebug'
+    AUTHORS = ['Ethan Reesor']
+    WEBSITE = 'https://gitlab.com/firelizzard/byebug-dap'
+    LICENSE = 'Apache-2.0'
+  end
+end